requirements.txt
minicli
icalendar
python-slugify
